# apispecs/src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Web Project API
  description: Api connect to Database
  version: 0.0.1
  license:
    name: TDTU Web Project
servers:
  - url: http://XXXXX/v1
    description: Develop server
  - url: http://XXXXX.com
    description: Product server
paths:
...
  /v1/semester:
    get:
      tags:
        - Semester
      summary: Api for get Semester
      operationId: semester
      parameters:
        - name: semesters
          in: query
          schema:
            type: string
            pattern: '^(\d{1,19},)*\d{1,19}$'
            description: Semester
            example: '200'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Semester
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemesterResponse'
...
  /v1/subject:
    get:
      tags:
        - Subject
      summary: Api for get Subject
      operationId: Subject
      parameters:
        - name: subjectIds
          in: query
          schema:
            type: string
            description: subjectIds
            example: '200'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Semester
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
...
  /v1/subjectGroup:
    get:
      tags:
        - Subject
      summary: Api for get Subject Group
      operationId: subjectGroup
      parameters:
        - name: groupIds
          in: query
          schema:
            type: string
            description: subjectIds
            example: '200'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Subject Group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectGroupResponse'
  /v1/createSubject:
    post:
      tags:
        - Subject
      summary: Api for create Subject
      operationId: createSubject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Create Training Process
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectCreateResponse'
  /v1/deleteSubject:
    delete:
      tags:
        - Subject
      summary: Api for delete Subject
      operationId: deleteSubject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectDeleteRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectDeleteResponse'
  /v1/updateSubject:
    put:
      tags:
        - Subject
      summary: Api for update Subject
      operationId: updateSubject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectUpdateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectUpdateResponse'
  /v1/universityStandards:
    get:
      tags:
        - LecturerStandard
      summary: Api for get University Lecturer Standards
      operationId: universityStandards
      parameters:
        - name: semester
          required: true
          in: query
          schema:
            type: number
            description: Semester for get Lecturer Standards
            example: 200
            minimum: 190001
            maximum: 299903
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Lecturer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UniversityStandardsResponse'
  /v1/masterStandards:
    get:
      tags:
        - LecturerStandard
      summary: Api for get Master Lecturer Standards
      operationId: masterStandards
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Lecturer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterStandardsResponse'
...
  /v1/subjectTrainingSys:
    get:
      tags:
        - Subject
      summary: Api for get Subject Training System
      operationId: subjectTrainingSys
      parameters:
        - name: trainingSysIds
          in: query
          schema:
            type: string
            description: trainingSysIds
            example: '200'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Subject Training System
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectTrainingSysResponse'
...
  /v1/teachingAssignment:
    post:
      tags:
        - Assignment
      summary: Api for Teaching Assignment
      operationId: teachingAssignment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Create Teaching Assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
  /v1/uploadLectureTime:
    post:
      tags:
        - Assignment
      summary: Api for Teaching Assignment
      operationId: uploadLectureTime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadAssignmentRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Create Teaching Assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadAssignmentResponse'
...
  /v1/teachingDiary:
    get:
      tags:
        - TeachingDiary
      summary: Api for get teaching Diary
      operationId: teachingDiary
      parameters:
        - name: historyId
          required: true
          in: query
          schema:
            type: string
            pattern: '^(\d{1,19},)*\d{1,19}$'
            description: List of ID Teaching Diary
            example: '200'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Teaching Diary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeachingDiaryResponse'
  /v1/updateTeachingDiary:
    put:
      tags:
        - TeachingDiary
      summary: Api for update Teaching Diary
      operationId: updateTeachingDiary
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeachingDiaryUpdateRequest'
      responses:
        400:
          $ref: '#/components/responses/webApiValidationError'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
        200:
          description: Teaching Diary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeachingDiaryUpdateResponse'
  #    End here
components:
...
  schemas:
    WebApiErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: array
          items:
            $ref: '#/components/schemas/WebApiErrorDetails'
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 400
        message:
          type: string

    #    Start here
...
    Subject:
      description: Subject
      type: object
      properties:
        maMon:
          type: string
          description: Ma mon of Subject
          example: 'maMon'
        phanLoai:
          type: string
          description: Phan Loai of Subject
          example: 'phanLoai'
        subjectGroup:
          $ref: '#/components/schemas/SubjectGroup'
        tenMon:
          type: string
          description: Subject
          example: 'tenMon'
        soTiet:
          type: number
          description: So tiet ly thuyet of Subject
          example: 200
    SubjectGroup:
      description: Subject group
      type: object
      properties:
        maNhom:
          type: string
          description: Ma nhom of Subject group
          example: 'maNhom'
        tenNhom:
          type: string
          description: Subject group
          example: 'tenNhom'
...
    SemesterDetailResponse:
      type: object
      properties:
        label:
          type: string
          description: Label of School year
          example: 'label'
        value:
          type: number
          description: Value of School year
          example: 200
        items:
          description: Semester
          type: array
          items:
            $ref: '#/components/schemas/Semester'
...
    SemesterResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            semesters:
              description: Semester list
              type: array
              items:
                $ref: '#/components/schemas/SemesterDetailResponse'
...
    SubjectResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            subjects:
              description: Subject list
              type: array
              items:
                $ref: '#/components/schemas/Subject'
            resultsTotalCount:
              description: Number of Subject with condition
              type: integer
              format: int64
              example: 1000
...
    SubjectGroupResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            subjectGroups:
              description: Subject Group  list
              type: array
              items:
                $ref: '#/components/schemas/SubjectGroup'
            resultsTotalCount:
              description: Number of Subject Group with condition
              type: integer
              format: int64
              example: 1000
    SubjectCreateRequest:
      type: object
      required:
        - subjectCreate
        - createBy
      properties:
        subjectCreate:
          $ref: '#/components/schemas/SubjectCreate'
        createBy:
          type: string
          description: Create By of Subject
          example: 'createBy'
    SubjectCreate:
      description: New Subject
      type: object
      required:
        - maNhom
        - maLoai
        - maMon
        - tenMon
        - soTiet
      properties:
        maNhom:
          type: string
          description: Ma nhom of Subject
          example: 'maNhom'
        maLoai:
          type: number
          description: Ma loai of Subject
          example: 200
        maMon:
          type: string
          description: Ma mon of Subject
          example: 'maMon'
        tenMon:
          type: string
          description: Ten mon of Subject
          example: 'tenMon'
        soTiet:
          type: number
          description: So tiet of Subject
          example: 200
          minimum: 0
          maximum: 360
    SubjectCreateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the create Subject
          type: string
          example: 'OK'
    SubjectDeleteRequest:
      type: object
      required:
        - subjectIds
        - deleteBy
      properties:
        subjectIds:
          type: string
          description: List ID of Subject
          example: '200'
        deleteBy:
          type: string
          description: Deleted By of Subject
          example: 'deleteBy'
    SubjectDeleteResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the delete Subject
          type: string
          example: 'OK'
    SubjectUpdateRequest:
      type: object
      required:
        - subjectId
        - subjectUpdate
        - updateBy
      properties:
        subjectId:
          type: string
          description: ID of Subject
          example: 'updateBy'
        subjectUpdate:
          $ref: '#/components/schemas/SubjectUpdate'
        updateBy:
          type: string
          description: Update By of Subject
          example: 'updateBy'
    SubjectUpdate:
      description: Update Subject
      type: object
      required:
        - tenMon
        - soTiet
      properties:
        tenMon:
          type: string
          description: soTiet of Subject
          example: 'soTiet'
        soTiet:
          type: number
          description: soTiet of Subject
          example: 200
          minimum: 0
          maximum: 360
    SubjectUpdateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the update Subject
          type: string
          example: 'OK'
    UniversityStandardsResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            universityStandards:
              description: University Lecturer Standards list
              type: array
              items:
                $ref: '#/components/schemas/UniversityStandardDetailResponse'
...
    MasterStandardsResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            masterStandards:
              description: Master Lecturer Standards list
              type: array
              items:
                $ref: '#/components/schemas/MasterStandardDetailResponse'
...
    SubjectTrainingSysResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            trainingSys:
              description: Subject Training Sys list
              type: array
              items:
                $ref: '#/components/schemas/TrainingSys'
            resultsTotalCount:
              description: Number of Subject Training Sys with condition
              type: integer
              format: int64
              example: 1000
    TrainingSys:
      description: Training Sys
      type: object
      properties:
        maHe:
          type: number
          description: maHe of Training Sys
          example: 200
        phanHe:
          type: string
          description: phanHe of Training Sys
          example: 'phanHe'
...
    AssignmentRequest:
      type: object
      required:
        - assignment
        - createBy
      properties:
        assignment:
          $ref: '#/components/schemas/Assignment'
        createBy:
          type: string
          description: Create By of Assignment
          example: 'createBy'
    UploadAssignmentRequest:
      type: object
      required:
        - assignments
        - createBy
      properties:
        assignments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Assignment'
        createBy:
          type: string
          description: Create By of Assignment
          example: 'createBy'
...
    AssignmentResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the create Assignment
          type: string
          example: 'OK'
    UploadAssignmentResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the create Assignment
          type: string
          example: 'OK'
...
    TeachingDiaryResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        results:
          type: object
          properties:
            teachingDiary:
              description: Teaching Diary list
              type: array
              items:
                $ref: '#/components/schemas/TeachingDiaryDetailResponse'
...
    TeachingDiaryUpdateRequest:
      type: object
      required:
        - historyId
        - teachingDiaryUpdate
        - updateBy
      properties:
        historyId:
          type: string
          description: History Id of Teaching Diary
          example: 200
        teachingDiaryUpdate:
          $ref: '#/components/schemas/TeachingDiaryUpdate'
        updateBy:
          type: string
          description: Update By of Teaching Diary
          example: 'updateBy'
...
    TeachingDiaryUpdateResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
          example: 200
        message:
          description: Message of the update Teaching Diary
          type: string
          example: 'OK'
    #    End here
  responses:
    webApiValidationError:
      description: webApi input value validation failure
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/WebApiErrorResponse'
            properties:
              status:
                example: 400
...
    401:
      description: Restrict access
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 401
              error:
                example: Unauthorized
    403:
      description: Access request denied
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 403
              error:
                example: Forbidden
    404:
      description: Page not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 404
              error:
                example: Not Found
    500:
      description: The server is having problems
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            properties:
              status:
                example: 500
              error:
                example: Internal Server Error
...